Separate the SharedVueStuff to 2 files, one for shared items between kitchen and ordering, and the other is between all user pages.
